requirements.txt
Keras_Preprocessing==1.1.2
matplotlib==3.5.1
numpy==1.22.0
pandas==1.3.5
Pillow==9.0.0
requests==2.27.1
scikit_learn==1.0.2
streamlit==1.3.1
tensorflow==2.7.0
wordcloud==1.8.1
